Re-organize some importing modules
