+++ matchms/importing/__init__.py
@@ -19,6 +19,7 @@ from .load_from_mgf import load_from_mgf
 from .load_from_msp import load_from_msp
 from .load_from_mzml import load_from_mzml
 from .load_from_mzxml import load_from_mzxml
+from .load_from_pickle import load_from_pickle
 from .load_from_usi import load_from_usi
 from .load_scores import scores_from_json, scores_from_pickle
 from .load_spectra import load_from_pickle, load_spectra
