+++ matchms/importing/load_from_mzml.py
@@ -1,7 +1,7 @@
 from typing import Generator
 import numpy as np
-from pyteomics import mzml
-from matchms.importing.parsing_utils import parse_mzml_mzxml_metadata
+from pyteomics.mzml import read
+from matchms.importing.parsing_utils import parse_mzml_mzxml_metadata, sort_spectrum
 from matchms.Spectrum import Spectrum
 
 
@@ -32,18 +32,14 @@ def load_from_mzml(filename: str, ms_level: int = 2,
         Set to False if metadata harmonization to default keys is not desired.
         The default is True.
     """
-    for pyteomics_spectrum in mzml.read(filename, dtype=dict):
-        if "ms level" in pyteomics_spectrum and pyteomics_spectrum["ms level"] == ms_level:
-            metadata = parse_mzml_mzxml_metadata(pyteomics_spectrum)
-            mz = np.asarray(pyteomics_spectrum["m/z array"], dtype="float")
-            intensities = np.asarray(pyteomics_spectrum["intensity array"], dtype="float")
+    with read(filename, dtype=dict) as reader:
+        for pyteomics_spectrum in reader:
+            if "ms level" in pyteomics_spectrum and pyteomics_spectrum["ms level"] == ms_level:
+                metadata = parse_mzml_mzxml_metadata(pyteomics_spectrum)
+                mz = np.asarray(pyteomics_spectrum["m/z array"], dtype="float")
+                intensities = np.asarray(pyteomics_spectrum["intensity array"], dtype="float")
 
-            if mz.shape[0] > 0:
-                # Sort by mz (if not sorted already)
-                if not np.all(mz[:-1] <= mz[1:]):
-                    idx_sorted = np.argsort(mz)
-                    mz = mz[idx_sorted]
-                    intensities = intensities[idx_sorted]
+                mz, intensities = sort_spectrum(mz=mz, intensities=intensities)
 
                 yield Spectrum(mz=mz, intensities=intensities, metadata=metadata,
                                metadata_harmonization=metadata_harmonization)
